SPSSWriter: Fix value of $sysmis.
$sysmis is supposed to be a floating-point value that, when written to an
SPSS system file, represents the system-missing value.  The previous value
was incorrect: when written to a system file, it yielded the bytes
fe ff ff ff ff ff ef c3, instead of the correct bytes
ff ff ff ff ff ff ef ff.  This was because of the difference between
integer and floating point: when -0xFFFFFFFFFFFFEFFF is converted to an
IEEE-754 "double", the bit representation changes.  This commit corrects
the problem, as can be observed by running test_write.php and observing
the bytes written in the machine floating point record at offset 0xb40.